snapshot: move API related code in to _api.go packages
